blockchain/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.6
black==23.7.0
Brownie==0.5.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cytoolz==0.12.1
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.8.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
parsimonious==0.9.0
pathspec==0.11.1
platformdirs==3.9.1
protobuf==4.23.4
pycryptodome==3.18.0
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.9.2
semantic-version==2.10.0
six==1.16.0
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.4
vyper==0.3.9
web3==6.6.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
